Add and edit flexbox features
Change the display of various elements to "flex" to align various
elements in particularly ways. Change margin within the header-container
 and add flex values to the header-container, content, and footer divs
to position them in such a way that more then half the webpage is taken
up by content, the header-container is slightly smaller than the content
and the footer takes little space at the bottom of the page.