Remove body height and specify container heights
Removed "height: 100vh" from the body's style and the flex values from
the child containers (header-container, content, footer) so that the
document elements could be larger and better match the proportions of
given full design of the project. Instead of using flex values, height
values of the child containers were specified.